Move extension helper functions also to extensions module
modified:   src/flask_uploads/extensions.py
modified:   src/flask_uploads/flask_uploads.py
