+++ src/flask_uploads/extensions.py
@@ -1,5 +1,6 @@
 
 """Extension presets and extension configuration."""
+import os
 
 # This contains archive and compression formats (.gz, .bz2, .zip, .tar,
 # .tgz, .txz, and .7z).
@@ -93,3 +94,32 @@ class AllExcept(object):
 
     def __contains__(self, item):
         return item not in self.items
+
+
+def extension(filename):
+    ext = os.path.splitext(filename)[1]
+    if ext.startswith('.'):
+        # os.path.splitext retains . separator
+        ext = ext[1:]
+    return ext
+
+
+def lowercase_ext(filename):
+    """
+    This is a helper used by UploadSet.save to provide lowercase extensions for
+    all processed files, to compare with configured extensions in the same
+    case.
+
+    .. versionchanged:: 0.1.4
+       Filenames without extensions are no longer lowercased, only the
+       extension is returned in lowercase, if an extension exists.
+
+    :param filename: The filename to ensure has a lowercase extension.
+    """
+    if '.' in filename:
+        main, ext = os.path.splitext(filename)
+        return main + ext.lower()
+    # For consistency with os.path.splitext,
+    # do not treat a filename without an extension as an extension.
+    # That is, do not return filename.lower().
+    return filename
